requirements.txt
torch~=2.4
lightning~=2.4
torchmetrics
torchvision
einops
numpy
scipy
scikit-learn
pandas
matplotlib
wandb~=0.18
tqdm
rich
ipython
ipdb
jupyterlab
ipympl
timm
vertexai
opencv-python
mediapipe
Pillow
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
ultralytics
diffusers
transformers
accelerate
pyrealsense2